Add some structure, test for breaking imports, move helper functions away from test files, define make test to run the tests + added test coverage reports
